Add files via upload
# Lab Assignment 9: n-armed bandit

This repository contains the code for Lab Assignment 9, which focuses on understanding the concept of Exploitation vs Exploration in a simple n-arm bandit reinforcement learning task using the epsilon-greedy algorithm.

## Problem Statement

The problem involves a binary bandit with two rewards {1-success, 0-failure}. The bandit returns 1 or 0 for the action that we select, i.e., 1 or 2. The rewards are stochastic but stationary. We use an epsilon-greedy algorithm to decide upon the action to take for maximizing the expected reward.

We also develop a 10-armed bandit in which all ten mean-rewards start out equal and then take independent random walks (by adding a normally distributed increment with mean zero and standard deviation 0.01 to all mean-rewards on each time step). 
The 10-armed bandit is difficult to crack with a standard epsilon-greedy algorithm since the rewards are non-stationary. We modify the epsilon-greedy agent to track non-stationary rewards and show whether it is able to latch onto correct actions or not.